requirements.txt
sqlalchemy==2.0.23
pillow==10.1.0
python-docx==1.1.0
reportlab==4.0.7
openpyxl==3.1.2
python-dateutil==2.8.2
tkcalendar==1.6.1
psutil==5.9.6
flask==3.0.0
qrcode==7.4.2
requests==2.31.0
gunicorn==21.2.0